Use state machine to control multiplayer lobby. Only re-render on main thread.
Modelled on the flux-like uni-directional data flow. The view is only
re-rendered on the main thread, and only in response to a state change.
Only certain state-changes are valid.

This is a work in progress - it has removed all notion of player lists
and avatars, to be added back next.