Only update end-game multiplayer screen on main thread.
Receiving network events means updating the UI to communicate these to the player.
Doing so in a background thread is problematic, as if Scene2D is mid-way through a render
of the screen, and then we start changing the content of labels, things go wrong. For example,
it is very common for scene2d to have measured a text layout, only for it to then be changed
prior to actually looping over the glyphs for rendering, at which time we end up with all
sorts of weird and wonderful NPE.

The solution to this is to only ever mark the screen as dirty from a network thread.

The main thread will check if it is dirty each render, and if so, rebuild the UI on the main
thread.