Websockets implementation of client + server.
Only thing it is lacking is the ability to discover other servers on the network.
Will probably need to use something like jmDNS.

The JSON messages sent via websockets use single character key names to make the
messages shorter. Is this worth it? Not sure, but doesn't really hurt as they can
still be read (and the binary Kryonet ones couldn't be read easily anyway).