Offer ability to only run non-android projects on Heroku.
New gradle plugins requrie an SDK just to configure the android
project, let alone actually run it. This adds a flag -PonlyServer
which ensures that the Android (and other peripheral subprojects)
are excluded from the gradle configuration step.

So instead, we just use multidex so we can have more than 65k classes.