Tempest: Add tempest icon.
To add the icon, we first add it to the global `images/icons/icons.svg` to make
sure it has a similar stil as other icons.

Then, from Inkscape, it is exported to an individual icon `images/icons/icon_tempest.png`.

We manually copy it to `texture-packer/sprites/icon_tempest.png` so that it is
available to the texture packer, then we run `./gradlew :packTextures to make available`
to add it to the texture atlas (which is automatically written to `android/assets/sprites.*`.

Finally, in Kotlin we add it to the list of icons that are available in `UiAssets` so it
can be used.