Tempest: Add empty game with controller.
This is the bare minimum required in order to add a new game
to retrowars (assumping it uses a soft controller and not a touch
screen - if just touch screen such as Missile Command then we
could ommit the controller).

Not particularly happy with the number of places things need
to be added, but that can be addressed in the future as we add
more games and realise where refactorings would be most appropriate.