Correctly cancel find-public-servers job when a server is selected.
The UI correctly allows the user to join a server before all other servers
are found. This is so that if there is a slow server taking time to respond
(or indeed timing out because it is no longer up).

However, this also means that if you select a server, then after you are
sitting in the lobby the view will wish to update with a new server, which
is an illegal state transition.

To prevent this issue, we record the `Job` being used for the background
coroutines to fetch server metadata, and cancel it when selecting a server.