More robust handling of end-game scenarios when the server tells us who won.
Client who won would accidentally echo back a status change of dead to the server after
already in final scores. This in turn triggered a secondary request to end the game.

Furthermore, it only properly checked for scenarios where one player was still playing
and they had the highest score. It neglected to also check for the scenario where all
players had died. In the past this didn't matter because that meant all players were
already in the end game screen. However now the server needs to know this because it
schedules a task to advance the game back to the lobby.