Players voluntarily leave the server when the screen is paused in Android.
After a lot of thought about how best to tell the server that a player
is still interested in playing, I've instead opted to just be pretty
harsh and not let players even lock their screens when connected to
the server. This will trigger an error that explains to the user the
game must be left open and to please rejoin to continue playing.

Also improved error messages slightly, and also made the server
properly send a shutdown message to each client when the JVM is
shutdown. This results in an appropriate message being shown to the
players who are connected.