Show a more comprehensive end game screen.
By recording the last survivor, we are able to properly inform them
as to why they were kicked to the last screen.

Specifically: When you have the highest score, and the second last player dies leaving
you as sole survivor - then one of three things will happen:

 1. You die before reaching the best score. This is fine and normal, and doesn't feel weird.
 2. You already had the highest score before the second last person died.
 3. You keep scoring after everyone else is dead until you become the highest scorer.

The final two force you to be kicked to the final score screen so that the game doesn't
drag on pointlessly when we know who the winner is. This is good for all other players
waiting for the next game to start, but feels weird for the sole survivor because they
don't really know why they went from happily playing to randomly thrown to the final score
screen.