Tempest: More tempest-like movement and crawling. Better rendering of enemies.
Allow crawling *before* reaching the end of the screen. Intersperse periods of crawling
with periods of walking straight. Sometimes change direction.

If we are at the end of the screen, crawl at a continual rate in the same direction.

Rendering now includes flatter enemies, which flip from one segment to the next.