Tempest: Enemies fire. Less score for tankers. Better collision detection.
Enemies fire at the start of their trip. Bullet's that they've fired don't
contribute to your score when shot - they are just a nuisance.

Tankers score less than flippers as per the original game (though the scores
don't line up with the original - we need to play test this to see how
it lines up with other games in multiplayer).

Collision detection was accidentally failing during the continuous phase.
It was checking the wrong direction (e.g. < in stead of >).
This has been rectified.