Tempest: Fire and move rapidly by pressing and holding buttons.
Turns the game into less of a button mashing experience
that rewards fast pressing of buttons, and allows you
to think more about the strategic elements of the game.

This behaviour mimics the original rapid fire behaviour
in the arcade version where a maximum of 8 bullets can
be on the screen at once, bu tyou can unleash them very quickly.