Space Invaders: More intelligent enemy firing.
Before, the firing always came from the bottom most non-empty row. This was a problem
if there was only one enemy in this row, but many in the rows above - all the bullets
would continue to come from one (potentially very distant to the player) location.

This new change makes it so that any row is capable of firing at the player. If a row
is chosen to fire, then its lowest enemy is the one which fires.