Add music to games from the wonderful SpaceJacked Soundtrack.
Mute button added to main menu and in game menus. Considered adding it to the options
screen too, but for now will leave as is. Potentially in the future it could be left
in the corner of all menu screens.

Moved images -> assets as it now contains music (and will soon contain sound effects).

Pushed game messages down into the GameDetails class instead of GameScreen constructor.
This is to try to put all the game specific details in one place.