More informative summary of server activity. Bump for release v0.22.0.
Previously, only active games would show. Given the game is not super popular, these were
few and far between. By instead focussing on the last time a player *joined the server*
rather than the last time *a game was played*, players may be more likely to hang
around for a short period while others join the lobby.

Bump in version number required due to the way we serialize/deserialize server info.
If it remained the same, we wouldn't be able to differentiate ServerInfoDTO objects
between old and new servers.