Space invaders: Spawn enemies closer to the center of mass, rather than fanning outwards early.
Prior to this, there was a tendency for the lowest row to always respan right until the edges. This meant that even if you were doing very
well and had contained the flanks of the aliens, you would quickly revert back to a position of 'aliens are close to the edge, and so
will drop quickly'.

By instead preferring to fill in all the blank spaces within the flanks across all rows, before extending the bottom row out,
it feels more natural and hopefully gives the players a fighting chance.