Breakout: Measure everything off screen height, not width.
Previously, the height of bricks (and thus the space between bricks and paddles),
along with the speed of the ball, were all measured from the screen height.

This meant that on very wide screens, there is very little space to move before
the ball drops to the bottom. By instead measuring everything from the *height*
of the screen, we have more success in equalising the playing field.