Breakout: Subtle acceleration applied to paddle.
This was done to allow micro movements by short taps to the on screen buttons.
Without it, even little taps shoot the paddle across the screen. With this, the
time taken to accelerate to top speed is almost imperceptable, but that seems
enough to make it more controllable with small taps.

This also deccelerates the paddle the same way, which emperically seems to
make it easier to "land" the paddle in the right spot.

Finally, if traveling left at speed and the user taps right, it doesn't bother
to slow down, rather resetting the velocity to 0 and then starting again.