Revert "Merge pull request #127 from retrowars/bump-build-tools"
This reverts commit 2f78e54494c86e690139aab6d4d370a580f02e76, reversing
changes made to cf2b0c46ddf8d10f0cfc0778d7f7b35e615a8302.

The upgrade was originally done just because AndroidStudio recommended
it, not because it was needed for any reason. It ended up caused issues
where the gradle Java version and the build JDK was different (Gradle
needed Java 8 but libgdx needed 8). Not worth it at this point, so
reverting.