Visually show supported games for each server.
Each server entry now shows a list of icons showing
which games are supported.

The hostname is de-emphasised to a smaller font.

Also remove the prominent "Active a few hours ago" message unless you
tap the "Info" button. Retain the "N players" message if there is at
least one player on the server so that it is obvious where to go to
play a game if others are around.